Add some cli tests
