+++ packages/compiler/core/cli/index.ts
@@ -1,0 +1,1 @@
+export * from "./args.js";
